Move server.go to internal
